Feature/testalphav2 (#35)
* remove unnecessary file and code

* fixed comp left page

* fixed daily action reward error

* quest profile component added

* added quest profile component

* Connect Twitter component reverted back

* fix quest layout

* fix twitter bug

* fixed twitter bug

* fix connect twitter in case twitter linked & profile page

* integrated fetchLeaderboard function in leaderboard page to show top … (#24)

* integrated fetchLeaderboard function in leaderboard page to show top users

* currected minimum xcapx to unlock leaderboard screen

* wallet page integrated with fetchWallet function, routed to home screen after clicking on banner button, currected leaderboard slice starting index and installed dayjs lib to convert date to unix for sorting in wallet page

* login and signup with redirect implemented

* integrating link social

* Link Your Twitter - Profile Page fix.

* fixed header heading

* linkwithredirect and signinWithRedirect integeated, header heading fixed

* connect with google btn text fixed in profile page

* Conditional Linking

* added liskauthprovider flag to check lick provider in profile page

* getting quest order if from router instead of redux store in whole profile page

* removed unnecessary logs

* go to next quest bug fixed

* removed connect your instagram and google button from profile page

* update profile page functionality added in profile page

* discord authentication done

* environment variable added for discord auth

* endpoints fixed

* profile page fix

* profile page fix

* profile page fix

* loader added on signin and signup page

* delete button added on profile page

+++ src/hooks/useLinkAuthProviders.js
@@ -8,6 +8,7 @@ import {
   unlink,
 } from "firebase/auth";
 import { useState } from "react";
+import { useSelector } from "react-redux";
 import { auth } from "../firebase/firebase";
 
 export const useLinkAuthProviders = () => {
@@ -16,6 +17,7 @@ export const useLinkAuthProviders = () => {
   const [linkDone, setLinkDone] = useState(false);
   const [useAccessToken, setUseActionToken] = useState(null);
   const [socialRedirectProvider, setSocialRedirectProvider] = useState("");
+  const userData = useSelector((state) => state.user);
   const user = auth.currentUser;
 
   const linkWithSocail = async (method) => {
@@ -39,7 +41,7 @@ export const useLinkAuthProviders = () => {
       }
 
       try {
-        const tokenDetails = await auth.currentUser.getIdTokenResult();
+        const tokenDetails = await auth.currentUser.getIdTokenResult(true);
         if (method.toUpperCase() === "TWITTER") {
           if (!tokenDetails.claims?.firebase.identities["twitter.com"]) {
             await linkWithRedirect(user, provider);
@@ -86,7 +88,14 @@ export const useLinkAuthProviders = () => {
       }
 
       try {
-        await unlink(user, provider.providerId);
+        if (
+          user.providerData.length === 1 &&
+          userData.socials.discord_id.length === 0
+        ) {
+          throw new Error("You must have one social provider");
+        } else {
+          await unlink(user, provider.providerId);
+        }
       } catch (error) {
         console.log(error);
         setError(error);
